+++ src/test/resources/models/parseTest.yaml
@@ -273,7 +273,7 @@ components:
       in: query
       scheme: http
       bearerFormat: bearerFormat
-      flow:
+      flows:
         implicit:
           authorizationUrl: authorizationUrl
           tokenUrl: tokenUrl
